+++ Client/RabbitMqClient.php
@@ -81,21 +81,27 @@ class RabbitMqClient
      * @param string $queueName
      * @param AMQPMessage[] $messageList
      * @param int $delay
+     * @param array $deliveryTagContextList
      */
     public function rewindList(
         string $queueName,
         array $messageList,
-        int $delay = 0
+        int $delay = 0,
+        array $deliveryTagContextList = [],
     ): void {
         $this->ackList($messageList);
 
         foreach ($messageList as $message) {
+            $deliveryTag = (string) $message->getDeliveryTag();
+            $context = $deliveryTagContextList[$deliveryTag] ?? [];
+
             $headers = $message->has('application_headers') ? $message->get('application_headers') : new AMQPTable();
 
             $retryCount = $headers->getNativeData()[QueueHeaderOptionEnum::X_RETRY] ?? 0;
 
             $headers->set(QueueHeaderOptionEnum::X_DELAY, $delay * 1000);
             $headers->set(QueueHeaderOptionEnum::X_RETRY, ++$retryCount);
+            $headers->set(QueueHeaderOptionEnum::X_CONTEXT, $context);
 
             $message->set('application_headers', $headers);