+++ Exception/RewindPartialException.php
@@ -14,13 +14,15 @@ class RewindPartialException extends RuntimeException
      * @var int[]
      */
     private array $rewindDeliveryTagList;
+    private array $deliveryTagContextList;
 
     /**
      * @param int[] $rewindDeliveryTagList
+     * @param array $deliveryTagContextList
      *
      * @throws RabbitQueueException
      */
-    public function __construct(array $rewindDeliveryTagList)
+    public function __construct(array $rewindDeliveryTagList, array $deliveryTagContextList = [])
     {
         foreach ($rewindDeliveryTagList as $deliveryTag) {
             if (!is_int($deliveryTag)) {
@@ -29,6 +31,7 @@ class RewindPartialException extends RuntimeException
         }
 
         $this->rewindDeliveryTagList = $rewindDeliveryTagList;
+        $this->deliveryTagContextList = $deliveryTagContextList;
 
         parent::__construct('Consumer rewind partial message list');
     }
@@ -39,5 +42,10 @@ class RewindPartialException extends RuntimeException
     public function getRewindDeliveryTagList(): array
     {
         return $this->rewindDeliveryTagList;
+    }
+
+    public function getDeliveryTagContextList(): array
+    {
+        return $this->deliveryTagContextList;
     }
 }